# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Eventloom API
  version: 1.0.0
  summary: Organization-scoped application operations, webhook administration, and published projections.
  description: |
    Direct clients use the operator's configured API origin. Browser traffic
    may use the deployed web app's same-origin `/api/*` proxy; it forwards to
    the API Worker and is not a separate API origin.

    The checked-in document intentionally combines the broader internal application API paths
    (`/api/admin/...`, `/api/speaker/...`, and `/api/public/...`) with tenant-scoped public-v1
    webhook administration. Internal routes may require organizer or participant session
    authentication and are not public-v1 guarantees.

    Organizer routes are qualified by organizationId and eventId and require an authenticated
    organizer session. Participant routes require the authenticated account's event grant.
    Public-v1 webhook routes are API-key authenticated and tenant-bound by organizationId.
    Separate public embeds and feeds expose immutable published projections only; they never expose
    organizer drafts, private assets, recipient data, evaluation notes, advisory candidates, or
    internal storage coordinates.

    Generic public-v1 events, sessions, speakers, and agenda routes are intentionally not mounted.
    They require publication-safe field allowlists, bounded projection reads, and durable
...
    authenticated principal's grants.
  license:
    name: Elastic License 2.0
...
servers:
...
security: []
paths:
...
  /api/admin/organizations/{organizationId}/events/{eventId}:
    parameters:
      - $ref: "#/components/parameters/OrganizationId"
      - $ref: "#/components/parameters/EventId"
    get:
      tags: [Events]
      operationId: getEvent
      summary: Read one event
      security:
        - sessionAuth: []
      responses:
        "200":
          description: Event visible to the authenticated organization membership.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventResponse" }
        "401": { $ref: "#/components/responses/AuthenticationError" }
        "403": { $ref: "#/components/responses/AuthorizationError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "500": { $ref: "#/components/responses/InternalError" }
    patch:
      tags: [Events]
      operationId: updateEvent
      summary: Update an event with optimistic concurrency
      security:
        - sessionAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/EventUpdate" }
      responses:
        "200":
          description: Event updated.
          headers:
            X-Request-ID: { $ref: "#/components/headers/RequestId" }
            ETag: { $ref: "#/components/headers/ETag" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EventResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/AuthenticationError" }
        "403": { $ref: "#/components/responses/AuthorizationError" }
        "404": { $ref: "#/components/responses/NotFoundError" }
        "409": { $ref: "#/components/responses/ConflictError" }
        "412": { $ref: "#/components/responses/PreconditionError" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      bearerFormat: Better Auth session
      description: Verified account session presented through the Authorization bearer credential or Better Auth session cookie. Event and organization grants are resolved server-side; query strings never grant access.
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: Eventloom scoped API key
      description: Organization-scoped API key for mounted public-v1 webhook administration.
    capabilityToken:
      type: apiKey
      in: header
      name: X-Capability-Token
      description: Single-use capability bound to the asset and transfer request; it is not an account credential.
  parameters:
    OrganizationId:
      name: organizationId
      in: path
      required: true
      description: Organization tenant selected by the route; it must match the authenticated membership or API key.
      schema: { type: string, minLength: 1, maxLength: 200, example: ai-engineer }
...
    EventId:
      name: eventId
      in: path
      required: true
      description: Event selected inside the organization or account grant.
      schema: { type: string, minLength: 1, maxLength: 200 }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique key for one intended mutation; a different body with the same key is rejected.
      schema: { type: string, minLength: 8, maxLength: 128 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: Required header-only positive resource version from the latest ETag; a body expectedVersion value is not a substitute. Strong or weak quoted values are accepted.
      schema: { type: string, pattern: '^(W/)?"?[1-9][0-9]*"?$' }
...
  headers:
    RequestId:
      description: UUID request trace. The accepted or generated value is echoed on the response.
      schema: { type: string, format: uuid }
    ETag:
      description: Strong resource version used with If-Match.
      schema: { type: string, pattern: '^"[1-9][0-9]*"$' }
...
    PrivateCache:
      description: Protected responses are never stored by shared caches.
      schema: { type: string, const: 'private, no-store' }
...
  schemas:
...
    Identifier:
      type: string
      minLength: 1
      maxLength: 200
    Timestamp:
      type: string
      format: date-time
    Version:
      type: integer
      minimum: 1
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, traceId]
          properties:
            code:
              type: string
              enum: [AUTHENTICATION_REQUIRED, ACCESS_DENIED, TENANT_SCOPE_VIOLATION, NOT_FOUND, VALIDATION_FAILED, CONFLICT, PRECONDITION_FAILED, IDEMPOTENCY_CONFLICT, RATE_LIMITED, INTEGRATION_UNAVAILABLE, CONFIGURATION_ERROR, INTERNAL_ERROR]
            message: { type: string, minLength: 1 }
            traceId: { type: string, format: uuid }
            details:
              type: array
              items: { $ref: "#/components/schemas/ValidationIssue" }
            retryAfterSeconds: { type: integer, minimum: 1 }
          additionalProperties: false
      additionalProperties: false
    ValidationIssue:
      type: object
      required: [path, code, message]
      properties:
        path:
          type: array
          items: { type: [string, integer] }
        code: { type: string, minLength: 1 }
        message: { type: string, minLength: 1 }
      additionalProperties: false
    Event:
      type: object
      required: [id, organizationId, slug, name, status, timeZone, startsAt, endsAt, version]
      properties:
        id: { $ref: "#/components/schemas/Identifier" }
        organizationId: { $ref: "#/components/schemas/Identifier" }
        slug: { type: string, minLength: 1, maxLength: 80 }
        name: { type: string, minLength: 1, maxLength: 200 }
        status: { type: string, enum: [draft, active, archived] }
        timeZone: { type: string, minLength: 1, maxLength: 128 }
        startsAt: { $ref: "#/components/schemas/Timestamp" }
        endsAt: { $ref: "#/components/schemas/Timestamp" }
        venue: { type: [string, "null"], maxLength: 2_000 }
        version: { $ref: "#/components/schemas/Version" }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
      additionalProperties: false
...
    EventUpdate:
      type: object
      minProperties: 1
      properties:
        slug: { type: string, minLength: 1, maxLength: 80 }
        name: { type: string, minLength: 1, maxLength: 200 }
        status: { type: string, enum: [draft, active, archived] }
        timeZone: { type: string, minLength: 1, maxLength: 128 }
        startsAt: { $ref: "#/components/schemas/Timestamp" }
        endsAt: { $ref: "#/components/schemas/Timestamp" }
        venue: { type: [string, "null"], maxLength: 2_000 }
        cfpSettings: { $ref: "#/components/schemas/EventCfpSettings" }
        defaultCalendarSettings: { $ref: "#/components/schemas/CalendarSettings" }
      additionalProperties: false
...
    EventCfpSettings:
      type: object
      properties:
        enabled: { type: boolean }
        opensAt: { type: [string, "null"], format: date-time }
        closesAt: { type: [string, "null"], format: date-time }
      additionalProperties: false
    CalendarSettings:
      type: object
      properties:
        durationMinutes: { type: integer, minimum: 1, maximum: 1_440 }
        timeZone: { type: string, minLength: 1, maxLength: 128 }
        location: { type: [string, "null"], maxLength: 2_000 }
      additionalProperties: false
    EventResponse:
      type: object
      required: [data]
      properties:
        data: { $ref: "#/components/schemas/Event" }
      additionalProperties: false
...
  responses:
    ValidationError:
      description: Request validation failed.
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestId" }
        Cache-Control: { $ref: "#/components/headers/PrivateCache" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    AuthenticationError:
      description: Authentication is missing or invalid.
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestId" }
        Cache-Control: { $ref: "#/components/headers/PrivateCache" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    AuthorizationError:
      description: The principal lacks the organization, event, account, or scope grant.
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestId" }
        Cache-Control: { $ref: "#/components/headers/PrivateCache" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFoundError:
      description: Resource is absent or hidden from this principal.
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestId" }
        Cache-Control: { $ref: "#/components/headers/PrivateCache" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ConflictError:
      description: State, idempotency, or integration conflict.
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestId" }
        Cache-Control: { $ref: "#/components/headers/PrivateCache" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    PreconditionError:
      description: If-Match is missing, malformed, or stale.
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestId" }
        Cache-Control: { $ref: "#/components/headers/PrivateCache" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    InternalError:
      description: Safe internal error with trace ID and no implementation detail.
      headers:
        X-Request-ID: { $ref: "#/components/headers/RequestId" }
        Cache-Control: { $ref: "#/components/headers/PrivateCache" }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }